requirements.txt
PyPDF2==3.0.1
python-docx==1.2.0
deep-translator==1.11.4
requests==2.32.5
pathlib
beautifulsoup4==4.14.2
lxml==6.0.2
anthropic==0.73.0
pydantic==2.12.4